Updated Data type for Latitude and Longitudes.

+++ app/src/main/java/com/patsurvey/nudge/utils/LocationCoordinates.kt
@@ -1,6 +1,6 @@
 package com.patsurvey.nudge.utils
 
 data class LocationCoordinates(
-    val lat: Int? = null,
-    val long: Int? = null
+    val lat: Double? = null,
+    val long: Double? = null
 )
